pyproject.toml
[tool.poetry]
name = "para-stats"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.8.2"
requests = "^2.31.0"
sqlalchemy = "^2.0.25"
pandas = "^2.1.4"
numpy = "^1.26.3"
rich = "^13.7.0"
python-dotenv = "^1.0.0"
psycopg = "^3.1.16"

[tool.poetry.scripts]
run = "main:init_script"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
